Refactor the backend impl package - 1st round of fixes - move repository package content level up
